Infer names for anonymous functions memoized with useCallback
Summary:
Currently, Metro infers the name `useCallback$argument_0` for callback functions memoized with `useCallback`. Callbacks are almost always assigned to a named variable, so let's use that instead.

For simplicity, we assume that we're always dealing with idiomatic React code, where any callee of the form `useCallback` or `React.useCallback` *is* React's `useCallback`. The impact of false positives/negatives here is negligible.

Reviewed By: rh389

Differential Revision: D33427900

fbshipit-source-id: 6ebc16abe9a0528da4df0f760f3657a735b947b6